Restructure `src/` folder
