+++ src/App/App.js
@@ -5,12 +5,12 @@ import AppHeader from './AppHeader';
 import AppFooter from './AppFooter';
 
 // Sections
-import HomeSection from './HomeSection';
-import AboutSection from './AboutSection';
-import ResumeSection from './ResumeSection';
-import PortfolioSection from './PortfolioSection';
-import SkillSection from './SkillSection';
-import ContactSection from './ContactSection';
+import HomeSection from 'sections/HomeSection';
+import AboutSection from 'sections/AboutSection';
+import ResumeSection from 'sections/ResumeSection';
+import PortfolioSection from 'sections/PortfolioSection';
+import SkillSection from 'sections/SkillSection';
+import ContactSection from 'sections/ContactSection';
 
 // Normalize default stylesheets
 import 'normalize.css/normalize.css';
