--- src/AppHeader/index.js
@@ -1,3 +1,0 @@
-import AppHeader from './AppHeader';
-
-export default AppHeader;
